Add data-sbgl-navxml to generate XML files without HTML tags.
This experimental feature allows to generate a sitemap.xml using the nav tag.
It parses the Tag Symbols and works like a nav, but skip printing the
nav, ul, and li in the output.  It could eventually also be used for custom
non-Atom feeds using a template.

See https://reykfloeter.com/sitemap.xml and
https://reykfloeter.com/templates/sitemap.xml as an example.